Rewrite process in modules files

+++ src/Base/Classes/led.cs
@@ -1,7 +1,7 @@
 public static class Led{
 	public static void on(byte r , byte g, byte b) => bc.TurnLedOn(r, g, b);
 
-	public static void on(Color color) => bc.TurnLedOn(color.r, color.g, color.b);
+	public static void on(Color color) => bc.TurnLedOn((int)color.r, (int)color.g, (int)color.b);
 
 	public static void off() => bc.TurnLedOff();
 }
